+++ lib/rubocop/rspec/config_formatter.rb
@@ -5,6 +5,7 @@ module RuboCop
     # Builds a YAML config file from two config hashes
     class ConfigFormatter
       NAMESPACES = /^(#{Regexp.union('RSpec', 'FactoryGirl')})/
+      STYLE_GUIDE_BASE_URL = 'http://www.rubydoc.info/gems/rubocop-rspec/RuboCop/Cop/'.freeze
 
       def initialize(config, descriptions)
         @config       = config
@@ -19,7 +20,9 @@ module RuboCop
 
       def unified_config
         cops.each_with_object(config.dup) do |cop, unified|
-          unified[cop] = config.fetch(cop).merge(descriptions.fetch(cop))
+          unified[cop] = config.fetch(cop)
+            .merge(descriptions.fetch(cop))
+            .merge('StyleGuide' => STYLE_GUIDE_BASE_URL + cop)
         end
       end