+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add `RSpec/MultipleSubjects` cop. ([@backus][])
 * Add `RSpec/ReturnFromStub` cop. ([@Darhazer][])
 * Add `RSpec/VoidExpect` cop. ([@pocke][])
+* Change HookArgument cop to detect when hook has a receiver. ([@pocke][])
 
 ## 1.15.1 (2017-04-30)